some documentation

+++ lib/jebson.php
@@ -1,6 +1,12 @@
 <?php
-error_reporting(E_ALL);
-ini_set('display_errors', '1');
+/**
+ * Jebson
+ * A pseudo static CMS
+ * @author Garrett Grimm
+ * @date September, 2013
+ *
+ */
+
 require_once 'lib/yaml.php';
 class Jebson {
 	// Setup
@@ -10,6 +16,7 @@ class Jebson {
 	public static $blogURI = 'blog';
 	public static $postsPerPage = 5;
 	public static $cache = false;
+	public static $debug = true;
 	
 	// Instance data
 	public static $request;
@@ -30,24 +37,43 @@ class Jebson {
 						'keywords'		=>''
 			  		);
 
+	/**
+	 * Main function that puts everything together.  The only method that's called from index.php
+	 * @return void
+	 */
 	public static function init() {
+		if (self::$debug) {
+			error_reporting(E_ALL);
+			ini_set('display_errors', '1');
+		}
+		
 		self::$startTime = microtime(true);
 		ob_start();
 		self::getParsedRequest();
 		self::getContent(implode('-', self::$request).'.html');
+		self::$loadTime = microtime(true) - self::$startTime;
 		self::buildPage();
 		
-		// If cache is enabled then save the completed page in the cache folder
-		if (self::$cache) {
+		// If cache is enabled and this page hasn't been cached then save the completed page in the cache folder
+		if (self::$cache && !file_exists('cache/'.str_replace('/', '-', $_SERVER['REQUEST_URI']).'.html')) {
 			file_put_contents('cache/'.str_replace('/', '-', $_SERVER['REQUEST_URI']).'.html', ob_get_contents());
 		}
 		ob_flush();
 	}
 
+	/**
+	 * Parses request URI into an array and stores as self::$request
+	 * @return void
+	 */
 	public static function getParsedRequest() {
 		self::$request = array_values(array_filter(explode('/', $_SERVER['REQUEST_URI'])));
 	}
 
+	/**
+	 * Gets content from given filename and saves it to static properties
+	 * @return void
+	 *
+	 */
 	public static function getContent($filename = false) {
 		$postPath = self::$contentDirectory.$filename;
 		
@@ -65,12 +91,14 @@ class Jebson {
 			//self::$title = $parsedFilename['title'];
 			self::$date = $parsedFilename['date'];
 			self::$slug = $parsedFilename['slug'];
-			
-			return true;
 		}
-		return false;
 	}
 	
+	/**
+	 * Pulls Yaml from self::$content and saves it to static properties
+	 * @return void
+	 *
+	 */
 	public static function getYaml() {
 		self::$yaml['raw'] = Yaml::get(self::$content);
 		self::$yaml['parsed'] = Yaml::parse(self::$yaml['raw']);
@@ -88,22 +116,35 @@ class Jebson {
 		}
 	}
 	
+	/**
+	 * Includes each element of the page as defined by self::$templateLoadOrder
+	 * @return void
+	 */
 	public static function buildPage() {
-		self::$loadTime = microtime(true) - self::$startTime;
 		foreach (self::$templateLoadOrder as $template) {
 			include self::$templateDirectory.$template.'.php';
 		}
 	}
 	
+	/**
+	 * Outputs output to page (not from cache)
+	 * @return void
+	 *
+	 */
 	public static function renderContent() {		
 		if (self::$request[0] == self::$blogURI) {
 			self::$pageNumber = isset(self::$request[1]) && is_numeric(self::$request[1]) ? self::$request[1] : null;
 			
 			// List posts with excerpts here
-			foreach (self::getPosts(self::$pageNumber) as $post)
-			{
+			$posts = self::getPosts(self::$pageNumber);
+			if (count($posts)) {
+				foreach (self::getPosts(self::$pageNumber) as $post) {
 				self::getContent($post);
 				include self::$templateDirectory.'excerpt.php';
+				}
+			}
+			else {
+				self::error(404);
 			}
 		}
 		elseif (isset(self::$content)) {
@@ -114,9 +155,15 @@ class Jebson {
 		}
 	}
 	
+	/**
+	 * getPosts
+	 * @param int $page Page number
+	 * @return array
+	 */
 	public static function getPosts($page = null) {
 		if ($handle = opendir(self::$contentDirectory)) {
 			$postCount = 0;
+			$posts = array();
  			while (false !== ($entry = readdir($handle))) {
 				if (substr($entry, 0, 1) != '.' && is_numeric(substr(str_replace('-','',$entry), 0, 7))) {
 					$postCount++;
@@ -146,6 +193,11 @@ class Jebson {
 		}
 	}
 	
+	/**
+	 * parseFilename
+	 * @param string $filename Name of file to parse
+	 * @return array
+	 */
 	public static function parseFilename($filename) {
 		$return['date'] = substr($filename, 0, 10);
 		$return['raw_title'] = str_replace('.html', '', substr($filename, 11));
@@ -160,7 +212,6 @@ class Jebson {
 				header('HTTP/1.0 404 Not Found');
 				echo '<p>Sorry, this page does not exist.</p>';
 				break;
-			
 			default:
 				echo 'Error';
 				break;
